# backend/src/main/java/com/diploma/proforientation/scoring/demo/DemoController.java
package com.diploma.proforientation.scoring.demo;

import com.diploma.proforientation.dto.ml.ScoringResult;
import com.diploma.proforientation.scoring.llm.LlmScoringEngineImpl;
import com.diploma.proforientation.scoring.ml.impl.MlScoringEngineImpl;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import lombok.RequiredArgsConstructor;
import org.springframework.web.bind.annotation.*;

@RestController
@RequestMapping("/demo")
@RequiredArgsConstructor
public class DemoController {

    private final MlScoringEngineImpl mlEngine;
    private final LlmScoringEngineImpl llmEngine;

    @PostMapping("/ml")
    @Operation(
            summary = "Run ML scoring demo",
            description = """
                Executes the machine-learning scoring engine using a raw set of demo answers.
                This endpoint is intended for testing and demonstration purposes only.
                """
    )
    @ApiResponse(
            responseCode = "200",
            description = "ML scoring result successfully calculated",
            content = @Content(
                    mediaType = "application/json",
                    schema = @Schema(implementation = ScoringResult.class)
            )
    )
    public ScoringResult testMl(@RequestBody DemoRequest req) {
        return mlEngine.evaluateRaw(req.answers());
    }

    @PostMapping("/llm")
    @Operation(
            summary = "Run LLM scoring demo",
            description = """
                Executes the LLM-based scoring engine using a raw set of demo answers.
                Returns recommendations and explanations generated by the LLM.
                """
    )
    @ApiResponse(
            responseCode = "200",
            description = "LLM scoring result successfully generated",
            content = @Content(
                    mediaType = "application/json",
                    schema = @Schema(implementation = ScoringResult.class)
            )
    )
    public ScoringResult testLlm(@RequestBody DemoRequest req) {
        return llmEngine.evaluateRaw(req.answers());
    }
}