# documentation/api-docs/reference/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: "1.0.0"
  description: |
      API for profession orientation and quiz-based assessment.

      ## Versioning Strategy
      This API follows URI-based semantic versioning:
      - Major versions are encoded in the base path: /api/v1.0
      - Backward-incompatible changes trigger a new major version (v2.0)
      - Backward-compatible changes are released within the same version
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
- url: http://localhost:8082
  description: Generated server url
security:
- BearerAuth: []
paths:
...
  /traits/{id}:
    get:
      tags:
      - trait-controller
      summary: Get trait by ID
      description: Returns a single trait localized according to the current request
        locale
      operationId: getById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "404":
          description: Trait not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "200":
          description: Trait found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TraitDto"
    put:
      tags:
      - trait-controller
      summary: Update trait
      description: Updates an existing trait (ADMIN only)
      operationId: update_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTraitRequest"
        required: true
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "404":
          description: Trait not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "200":
          description: Trait updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TraitDto"
        "403":
          description: Access denied (ADMIN role required)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
    delete:
      tags:
      - trait-controller
      summary: Delete trait
      description: Deletes a trait by its ID (ADMIN only)
      operationId: delete_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "204":
          description: Trait deleted successfully
        "403":
          description: Access denied (ADMIN role required)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
...
  /traits:
    get:
      tags:
      - trait-controller
      summary: Get all traits
      description: Returns all traits localized according to the current request locale
      operationId: getAll
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "200":
          description: List of traits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TraitDto"
    post:
      tags:
      - trait-controller
      summary: Create trait
      description: Creates a new trait (ADMIN only)
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTraitRequest"
        required: true
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "400":
          description: Invalid request payload
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "200":
          description: Trait created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TraitDto"
        "403":
          description: Access denied (ADMIN role required)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
...
  /demo/ml:
    post:
      tags:
      - demo-controller
      summary: Run ML scoring demo
      description: |
        Executes the machine-learning scoring engine using a raw set of demo answers.
        This endpoint is intended for testing and demonstration purposes only.
      operationId: testMl
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemoRequest"
        required: true
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "200":
          description: ML scoring result successfully calculated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoringResult"
  /demo/llm:
    post:
      tags:
      - demo-controller
      summary: Run LLM scoring demo
      description: |
        Executes the LLM-based scoring engine using a raw set of demo answers.
        Returns recommendations and explanations generated by the LLM.
      operationId: testLlm
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemoRequest"
        required: true
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "200":
          description: LLM scoring result successfully generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoringResult"
...
components:
  schemas:
    ExceptionDto:
      type: object
      description: Standardized API error response returned for all handled exceptions
      properties:
        code:
          type: integer
          format: int32
          description: HTTP status code representing the error type
          examples:
          - 400
        time:
          type: string
          format: date-time
          description: Timestamp indicating when the error occurred (ISO-8601 format)
          examples:
          - 2025-01-15T14:32:10Z
        message:
          type: object
          description: |
            Error details. May be:
            - a string message (e.g. 'Invalid credentials')
            - a structured object containing validation errors
          examples:
          - Invalid email or password
          oneOf:
          - type: string
          - {}
...
    CreateTraitRequest:
      type: object
      description: Request payload for creating a psychological trait
      properties:
        code:
          type: string
          description: "Unique trait code (e.g. R, I, A)"
          examples:
          - R
        name:
          type: string
          description: Human-readable trait name
          examples:
          - Realistic
        description:
          type: string
          description: Description explaining the trait meaning
          examples:
          - "Preference for practical, hands-on activities"
        bipolarPairCode:
          type: string
          description: Code of the opposite trait in a bipolar pair
          examples:
          - S
    TraitDto:
      type: object
      description: Represents a psychological trait used in assessment scoring
      properties:
        id:
          type: integer
          format: int32
          description: Unique identifier of the trait
          examples:
          - 1
        code:
          type: string
          description: Short code representing the trait (e.g. RIASEC dimension)
          examples:
          - R
        name:
          type: string
          description: Human-readable name of the trait
          examples:
          - Realistic
        description:
          type: string
          description: Detailed description of the trait
          examples:
          - Preference for hands-on activities and practical problem solving
        bipolarPairCode:
          type: string
          description: Code of the opposing trait in a bipolar pair
          examples:
          - S
...
    DemoRequest:
      type: object
      properties:
        answers:
          type: array
          items:
            type: integer
            format: int32
    RecommendationDto:
      type: object
      description: Career recommendation generated by the scoring engine
      properties:
        professionId:
          type: integer
          format: int32
          description: Identifier of the recommended profession
          examples:
          - 42
        score:
          type: number
          description: Identifier of the recommended profession
          examples:
          - 42
        explanation:
          type: string
          description: Human-readable explanation of why this profession was recommended
          examples:
          - "Predicted as: Software Engineer"
    ScoringResult:
      type: object
      description: Final scoring result produced after evaluating a quiz attempt
      properties:
        traitScores:
          type: object
          additionalProperties:
            type: number
          description: |
            Calculated trait scores based on quiz answers.
            Each entry maps a trait profile to its aggregated score.
          examples:
          - |
            {
              "R": 12.5,
              "I": 9.0,
              "A": 7.5
            }
        recommendations:
          type: array
          description: List of profession recommendations generated by the ML scoring
            engine
          examples:
          - |
            [
              {
                "professionId": 42,
                "score": 0.82,
                "explanation": "Predicted as: Software Engineer"
              }
            ]
          items:
            $ref: "#/components/schemas/RecommendationDto"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT